Added Updated Files For Milestone 5
Updates Include: Proper Updated Quantities when removing or updating items in the shopping cart so that the proper quantity is reflected in the Inventory ArrayList. Generics were added to Inventory and ShoppingCart Classes. Methods used for sorting were added to StoreFront, Inventroy, and SalableProduct Classes. StoreDriver has new piece of menu for user to select how they wish the inventory to be sorted.