Added StoreServer and Client Class Files
The StoreServer File communicates with the Client class to send and receive commands. The Update and Read commands (U and R) are supported from the client. All interaction with the store front is server side. With this update the StoreDriver Class is no longer needed, the StoreServer class acts as it essentially